Update Javascript Schedule Testing
This update has the gradient in the background of the schedule. It took me a long time to figure out why the gradient background was applying to everything in the canvas (so it would show up as just the gradient) but this solves everything. I have the colors set right now that it contrasts really well, but I'm not really sure what it'll look like when it's on the jsp page (that has a gray background). Either way, those colors are easily changeable. Run it and lemme know what you guys think!