Update Javascript Schedule Testing
This update gives each of the classes on the schedule a thick black outline. Also, this update will give kind of a radial gradient to the individual course blocks so they aren't just a boring solid red color anymore. I also added another few classes to demonstrate, and changed one of the meeting times to a half hour increment instead of beginning and ending on the hour. The only problem I'm coming across now is going to be drawing classes that are not a solid hour. If anyone wants to take a look into that for me, that'd be great but I'll keep working anyway.