Update Javascript Schedule Testing
In this update, I found out what why classes would only be drawn for an hour, even if I had it listed for 2. Just a simple logic error. Also in this update, I have the monday part of the method done. Once we get that working, all we have to do is copy and paste the code for the other days of the week and we should be all set.